+++ clan.events.client/src/app/pages/events/modules/board-canvas/board-canvas.component.scss
@@ -1,8 +1,6 @@
-:host {
-  #board-canvas {
-    border: 1px solid var(--app-border-color-z-3);
-    background-color: var(--app-background-color-z-3);
-    width: 100%;
-    min-height: 600px;
-  }
+#board-canvas {
+  border: 1px solid var(--app-border-color-z-3);
+  background-color: var(--app-background-color-z-3);
+  width: 100%;
+  min-height: 600px;
 }
