+++ clan.events.client/src/app/pages/events/modules/board/board.component.scss
@@ -1,47 +1,45 @@
-:host {
-  mat-spinner {
-    margin: 50px auto;
+mat-spinner {
+  margin: 50px auto;
+}
+
+.board-container {
+  display: flex;
+  gap: 24px;
+
+  .board-canvas-container {
+    position: relative;
+    flex: 1;
   }
 
-  .board-container {
+  .spinner-overlay {
+    position: absolute;
     display: flex;
-    gap: 24px;
+    align-items: center;
+    justify-content: center;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
+    background-color: var(--app-background-color-z-3);
+    opacity: 0.9;
+  }
 
-    .board-canvas-container {
-      position: relative;
-      flex: 1;
-    }
+  app-edit-tile {
+    width: 300px;
+  }
 
-    .spinner-overlay {
-      position: absolute;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      width: 100%;
-      height: 100%;
-      background-color: var(--app-background-color-z-3);
-      opacity: 0.9;
+  /* break at 1200px and put edit tile below board */
+  @media (max-width: 1200px) {
+    flex-direction: column;
+
+    app-board-canvas {
+      min-width: 0;
     }
 
     app-edit-tile {
-      width: 300px;
-    }
-
-    /* break at 1200px and put edit tile below board */
-    @media (max-width: 1200px) {
-      flex-direction: column;
-
-      app-board-canvas {
-        min-width: 0;
-      }
-
-      app-edit-tile {
-        width: 100%;
-      }
+      width: 100%;
     }
   }
 }
