+++ clan.events.client/src/app/pages/events/modules/board/renderers/board-canvas-object.ts
@@ -1,0 +1,6 @@
+export interface BoardCanvasObject {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
